created astra-sim/common and placed some common files, moved APIs to common and left links
